+++ src/ui/welcome.rs
@@ -1071,5 +1071,6 @@ pub(crate) fn blank_account() -> AccountConfig {
         empty_trash_days: 0,
         pgp_key: None,
         in_unified: true,
+        sign_by_default: false,
     }
 }